+++ packages/sdk/src/base.ts
@@ -204,6 +204,10 @@ export default class ConnectionBase {
    * Lyra インスタンス
    */
   private lyra?: LyraState
+  /**
+   * キーとなる sender が setupSenderTransform で初期化済みかどうか
+   */
+  private senderStreamInitialized: WeakSet<RTCRtpSender> = new WeakSet()
 
   constructor(
     signalingUrlCandidates: string | string[],
@@ -434,6 +438,7 @@ export default class ConnectionBase {
     }
     stream.addTrack(audioTrack)
     await transceiver.sender.replaceTrack(audioTrack)
+    await this.setupSenderTransform(transceiver.sender)
   }
 
   /**
@@ -465,6 +470,7 @@ export default class ConnectionBase {
     }
     stream.addTrack(videoTrack)
     await transceiver.sender.replaceTrack(videoTrack)
+    await this.setupSenderTransform(transceiver.sender)
   }
 
   /**
@@ -1441,6 +1447,10 @@ export default class ConnectionBase {
     if ((this.e2ee === null && this.lyra === undefined) || sender.track === null) {
       return
     }
+    // 既に初期化済み
+    if (this.senderStreamInitialized.has(sender)) {
+      return
+    }
 
     const isLyraCodec = sender.track.kind === 'audio' && this.options.audioCodecType === 'LYRA'
 
@@ -1484,6 +1494,7 @@ export default class ConnectionBase {
         readable.pipeTo(senderStreams.writable).catch((e) => console.warn(e))
       }
     }
+    this.senderStreamInitialized.add(sender)
   }
 
   /**