+++ CHANGES.md
@@ -29,6 +29,8 @@
     - @voluntas
 - [CHANGE] サンプルのチャネル ID を sora に変更する
     - @voluntas
+- [FIX] ミュート状態で接続すると、replace[Video|Audio]Track しても画像・音声データが送信されないのを修正
+    - @melpon
 
 ## 2022.3.1